feat 볼륨 스크롤 애니메이션 추가 및 슬라이더 커스텀
