+++ video-component/tailwind.config.js
@@ -6,6 +6,12 @@ module.exports = {
       colors: {
         "kakao-yellow": "#FAE100",
       },
+      spacing: {
+        12: "3rem",
+        32: "8rem",
+        48: "12rem",
+        64: "16rem",
+      },
     },
   },
   plugins: [],
